+++ e2e/node/e2e.test.ts
@@ -26,9 +26,13 @@ import {
   isValidAccessGrant,
   issueAccessRequest,
   revokeAccessGrant,
+  getFile,
 } from "../../src/index";
 import { Session } from "@inrupt/solid-client-authn-node";
-import { isVerifiableCredential } from "@inrupt/solid-client-vc";
+import {
+  isVerifiableCredential,
+  VerifiableCredential,
+} from "@inrupt/solid-client-vc";
 import { config } from "dotenv-flow";
 
 // Load environment variables from .env.test.local if available:
@@ -74,8 +78,11 @@ const serversUnderTest: AuthDetails[] = [
   ],
 ];
 
-const SHARED_IMAGE_IRI =
-  "https://storage.dev-next.inrupt.com/eb2f327b-7bb4-4ba2-9b4b-678a4d7e3551/adminadmin.jpg";
+const SHARED_FILE_IRI =
+  "https://storage.dev-next.inrupt.com/eb2f327b-7bb4-4ba2-9b4b-678a4d7e3551/somefile.txt";
+
+// This is the content of the file uploaded manually at SHARED_FILE_IRI.
+const SHARED_FILE_CONTENT = "Some content.\n";
 
 describe.each(serversUnderTest)(
   "Access grant client end-to-end tests authenticated to [%s], issuing from [%s] for [%s]",
@@ -114,6 +121,9 @@ describe.each(serversUnderTest)(
           oidcIssuer,
           clientId: requestorClientId,
           clientSecret: requestorClientSecret,
+          // Note that currently, using a Bearer token (as opposed to a DPoP one)
+          // is required for the UMA access token to be usable.
+          tokenType: "Bearer",
         });
         await resourceOwnerSession.login({
           oidcIssuer,
@@ -134,9 +144,8 @@ describe.each(serversUnderTest)(
           {
             access: { read: true },
             requestor,
-            requestorInboxUrl: "https://some.bogus.inbox",
             resourceOwner,
-            resources: [SHARED_IMAGE_IRI],
+            resources: [SHARED_FILE_IRI],
             purpose: [
               "https://some.purpose/not-a-nefarious-one/i-promise",
               "https://some.other.purpose/",
@@ -158,6 +167,7 @@ describe.each(serversUnderTest)(
             accessEndpoint: vcService,
           }
         );
+
         await expect(
           isValidAccessGrant(grant, {
             fetch: resourceOwnerSession.fetch,
@@ -169,16 +179,27 @@ describe.each(serversUnderTest)(
         ).resolves.toMatchObject({ errors: [] });
 
         const grantedAccess = await getAccessGrantAll(
-          SHARED_IMAGE_IRI,
+          SHARED_FILE_IRI,
           undefined,
           {
             fetch: resourceOwnerSession.fetch,
             accessEndpoint: vcService,
           }
         );
+
         // Test that looking up the access grants for the given resource returns
         // the access we just granted.
         expect(grantedAccess).toContainEqual(grant);
+
+        // For some reason, the Node jest runner throws an undefined error when
+        // calling to btoa. This overrides it, while keeping the actual code
+        // environment-agnostic.
+        global.btoa = (str: string) => Buffer.from(str).toString("base64");
+
+        const sharedFile = await getFile(SHARED_FILE_IRI, grant, {
+          fetch: requestorSession.fetch,
+        });
+        await expect(sharedFile.text()).resolves.toBe(SHARED_FILE_CONTENT);
 
         await revokeAccessGrant(grant, {
           fetch: resourceOwnerSession.fetch,
@@ -197,7 +218,7 @@ describe.each(serversUnderTest)(
       it("can filter VCs held by the service based on requestor", async () => {
         await expect(
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             { requestor },
             {
               fetch: resourceOwnerSession.fetch,
@@ -207,7 +228,7 @@ describe.each(serversUnderTest)(
         ).resolves.not.toHaveLength(0);
         await expect(
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             { requestor: "https://some.unknown.requestor" },
             {
               fetch: resourceOwnerSession.fetch,
@@ -219,7 +240,7 @@ describe.each(serversUnderTest)(
 
       it("can filter VCs held by the service based on target resource", async () => {
         await expect(
-          getAccessGrantAll(SHARED_IMAGE_IRI, undefined, {
+          getAccessGrantAll(SHARED_FILE_IRI, undefined, {
             fetch: resourceOwnerSession.fetch,
             accessEndpoint: vcService,
           })
@@ -240,12 +261,12 @@ describe.each(serversUnderTest)(
           bothPurposeFilter,
           unknownPurposeFilter,
         ] = await Promise.all([
-          getAccessGrantAll(SHARED_IMAGE_IRI, undefined, {
+          getAccessGrantAll(SHARED_FILE_IRI, undefined, {
             fetch: resourceOwnerSession.fetch,
             accessEndpoint: vcService,
           }),
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             { purpose: ["https://some.purpose/not-a-nefarious-one/i-promise"] },
             {
               fetch: resourceOwnerSession.fetch,
@@ -253,7 +274,7 @@ describe.each(serversUnderTest)(
             }
           ),
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             { purpose: ["https://some.other.purpose/"] },
             {
               fetch: resourceOwnerSession.fetch,
@@ -261,7 +282,7 @@ describe.each(serversUnderTest)(
             }
           ),
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             {
               purpose: [
                 "https://some.purpose/not-a-nefarious-one/i-promise",
@@ -274,7 +295,7 @@ describe.each(serversUnderTest)(
             }
           ),
           getAccessGrantAll(
-            SHARED_IMAGE_IRI,
+            SHARED_FILE_IRI,
             { purpose: ["https://some.unknown.purpose/"] },
             {
               fetch: resourceOwnerSession.fetch,