+++ jest.config.js
@@ -21,7 +21,8 @@
 
 module.exports = {
   preset: "ts-jest",
-  testEnvironment: "jsdom",
+  // this test environment adds TextEncoder to jsdom, which we require.
+  testEnvironment: "<rootDir>/tests/environment/customEnvironment.js",
   clearMocks: true,
   injectGlobals: false,
   collectCoverage: true,