+++ src/fetch/index.test.ts
@@ -1,0 +1,399 @@
+// Copyright 2021 Inrupt Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+// Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+// PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+// SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+// This rule complains about testing promises as described in the Jest docs:
+// https://jestjs.io/docs/asynchronous#promises
+/* eslint jest/no-conditional-expect: 0 */
+
+// Also disabling this to make my fetch mocking easy.
+/* eslint @typescript-eslint/no-explicit-any: 0 */
+
+// This rule complains about the `@jest/globals` variables overriding global vars:
+// eslint-disable-next-line no-shadow
+import { jest, describe, it, expect } from "@jest/globals";
+// eslint-disable-next-line no-shadow
+import { fetch as crossFetch, Response } from "cross-fetch";
+
+import {
+  parseUMAAuthTicket,
+  parseUMAAuthIri,
+  getUmaConfiguration,
+  exchangeTicketForAccessToken,
+  boundFetch,
+  fetchWithVc,
+} from "./index";
+
+jest.mock("cross-fetch", () => {
+  const crossFetchModule = jest.requireActual("cross-fetch") as any;
+  return {
+    ...crossFetchModule,
+    fetch: jest.fn(),
+  };
+});
+
+const MOCK_VC = {
+  "@context": [
+    "https://www.w3.org/2018/credentials/v1",
+    "https://consent.pod.inrupt.com/credentials/v1",
+  ],
+  id: "https://example.com/id",
+  issuer: "https://example.com/issuer",
+  type: ["VerifiableCredential", "SolidCredential", "SolidAccessGrant"],
+  issuanceDate: "2021-05-26T16:40:03",
+  expirationDate: "2021-06-09T16:40:03",
+  credentialSubject: {
+    id: "https://pod.inrupt.com/alice/profile/card#me",
+    providedConsent: {
+      mode: ["http://www.w3.org/ns/auth/acl#Read"],
+      hasStatus: "ConsentStatusExplicitlyGiven",
+      forPersonalData: "https://pod.inrupt.com/alice/private/data",
+      forPurpose: "https://example.com/SomeSpecificPurpose",
+      isProvidedToPerson: "https://pod.inrupt.com/bob/profile/card#me",
+    },
+  },
+  proof: {
+    created: "2021-05-26T16:40:03.009Z",
+    proofPurpose: "assertionMethod",
+    proofValue: "eqp8h_kL1DwJCpn65z-d1Arnysx6b11...jb8j0MxUCc1uDQ",
+    type: "Ed25519Signature2020",
+    verificationMethod: "https://consent.pod.inrupt.com/key/396f686b",
+  },
+};
+
+const mockFetch = (response: Response) => {
+  return (
+    crossFetch as jest.MockedFunction<typeof crossFetch>
+  ).mockResolvedValueOnce(response);
+};
+
+describe("parseUMAAuthTicket", () => {
+  it("parses and returns the ticket from a string", () => {
+    const header =
+      'UMA realm="test" as_uri="https://fake.url" ticket="some_value"';
+
+    const ticket = parseUMAAuthTicket(header);
+
+    expect(ticket).toEqual("some_value");
+  });
+
+  it("returns null if no value is found", () => {
+    const header = 'UMA realm="test" as_uri="https://fake.url"';
+
+    const ticket = parseUMAAuthTicket(header);
+    expect(ticket).toBeNull();
+  });
+});
+
+describe("parseUMAAuthIri", () => {
+  it("parses and returns the as_uri from a string", () => {
+    const header =
+      'UMA realm="test" as_uri="https://fake.url" ticket="some_value"';
+
+    const ticket = parseUMAAuthIri(header);
+    expect(ticket).toEqual("https://fake.url");
+  });
+
+  it("returns null if no value is found", () => {
+    const header = 'UMA realm="test" ticket="some_value"';
+
+    const ticket = parseUMAAuthIri(header);
+    expect(ticket).toBeNull();
+  });
+});
+
+describe("getUmaConfiguration", () => {
+  it("fetches an iri by appending the .well-known path", async () => {
+    const iri = "https://fake.url";
+    const expectedIri = "https://fake.url/.well-known/uma2-configuration";
+
+    const mockedFetch = mockFetch(new Response(JSON.stringify({})));
+
+    await getUmaConfiguration(iri);
+    expect(mockedFetch).toHaveBeenCalledWith(expectedIri);
+  });
+
+  it("throws if parsing the response as JSON fails", async () => {
+    const iri = "https://fake.url";
+    const expectedIri = "https://fake.url/.well-known/uma2-configuration";
+
+    mockFetch(new Response("Not a JSON document."));
+
+    await expect(getUmaConfiguration(iri)).rejects.toThrow(
+      `Parsing the UMA configuration found at ${expectedIri} failed`
+    );
+  });
+});
+
+describe("exchangeTicketForAccessToken", () => {
+  it("posts to the token endpoint and parses the resulting access token", async () => {
+    const tokenEndpoint = "https://fake.url/token";
+    const authTicket = "auth-ticket";
+
+    const mockedFetch = mockFetch(
+      new Response(
+        JSON.stringify({
+          access_token: "some_access_token",
+        })
+      )
+    );
+
+    const token = await exchangeTicketForAccessToken(
+      tokenEndpoint,
+      MOCK_VC,
+      authTicket,
+      mockedFetch
+    );
+
+    expect(token).toEqual("some_access_token");
+    expect(mockedFetch).toHaveBeenCalledWith(tokenEndpoint, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/x-www-form-urlencoded",
+      },
+      // The body should be the following data as a form_urlencoded:
+      // claim_token: btoa(JSON.stringify({ verifiableCredential: [MOCK_VC] })),
+      // claim_token_type: "https://www.w3.org/TR/vc-data-model/#json-ld",
+      // grant_type: "urn:ietf:params:oauth:grant-type:uma-ticket",
+      // ticket: authTicket,
+      body: "claim_token=eyJ2ZXJpZmlhYmxlQ3JlZGVudGlhbCI6W3siQGNvbnRleHQiOlsiaHR0cHM6Ly93d3cudzMub3JnLzIwMTgvY3JlZGVudGlhbHMvdjEiLCJodHRwczovL2NvbnNlbnQucG9kLmlucnVwdC5jb20vY3JlZGVudGlhbHMvdjEiXSwiaWQiOiJodHRwczovL2V4YW1wbGUuY29tL2lkIiwiaXNzdWVyIjoiaHR0cHM6Ly9leGFtcGxlLmNvbS9pc3N1ZXIiLCJ0eXBlIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiU29saWRDcmVkZW50aWFsIiwiU29saWRBY2Nlc3NHcmFudCJdLCJpc3N1YW5jZURhdGUiOiIyMDIxLTA1LTI2VDE2OjQwOjAzIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDIxLTA2LTA5VDE2OjQwOjAzIiwiY3JlZGVudGlhbFN1YmplY3QiOnsiaWQiOiJodHRwczovL3BvZC5pbnJ1cHQuY29tL2FsaWNlL3Byb2ZpbGUvY2FyZCNtZSIsInByb3ZpZGVkQ29uc2VudCI6eyJtb2RlIjpbImh0dHA6Ly93d3cudzMub3JnL25zL2F1dGgvYWNsI1JlYWQiXSwiaGFzU3RhdHVzIjoiQ29uc2VudFN0YXR1c0V4cGxpY2l0bHlHaXZlbiIsImZvclBlcnNvbmFsRGF0YSI6Imh0dHBzOi8vcG9kLmlucnVwdC5jb20vYWxpY2UvcHJpdmF0ZS9kYXRhIiwiZm9yUHVycG9zZSI6Imh0dHBzOi8vZXhhbXBsZS5jb20vU29tZVNwZWNpZmljUHVycG9zZSIsImlzUHJvdmlkZWRUb1BlcnNvbiI6Imh0dHBzOi8vcG9kLmlucnVwdC5jb20vYm9iL3Byb2ZpbGUvY2FyZCNtZSJ9fSwicHJvb2YiOnsiY3JlYXRlZCI6IjIwMjEtMDUtMjZUMTY6NDA6MDMuMDA5WiIsInByb29mUHVycG9zZSI6ImFzc2VydGlvbk1ldGhvZCIsInByb29mVmFsdWUiOiJlcXA4aF9rTDFEd0pDcG42NXotZDFBcm55c3g2YjExLi4uamI4ajBNeFVDYzF1RFEiLCJ0eXBlIjoiRWQyNTUxOVNpZ25hdHVyZTIwMjAiLCJ2ZXJpZmljYXRpb25NZXRob2QiOiJodHRwczovL2NvbnNlbnQucG9kLmlucnVwdC5jb20va2V5LzM5NmY2ODZiIn19XX0%3D&claim_token_type=https%3A%2F%2Fwww.w3.org%2FTR%2Fvc-data-model%2F%23json-ld&grant_type=urn%3Aietf%3Aparams%3Aoauth%3Agrant-type%3Auma-ticket&ticket=auth-ticket",
+    });
+  });
+
+  it("returns null if no body is returned", async () => {
+    const tokenEndpoint = "https://fake.url/token";
+    const authTicket = "auth-ticket";
+
+    const mockedFetch = mockFetch(new Response());
+
+    const token = await exchangeTicketForAccessToken(
+      tokenEndpoint,
+      MOCK_VC,
+      authTicket,
+      mockedFetch
+    );
+
+    expect(token).toBeNull();
+  });
+
+  it("returns null if no access token is returned", async () => {
+    const tokenEndpoint = "https://fake.url/token";
+    const authTicket = "auth-ticket";
+
+    const mockedFetch = mockFetch(new Response(JSON.stringify({})));
+
+    const token = await exchangeTicketForAccessToken(
+      tokenEndpoint,
+      MOCK_VC,
+      authTicket,
+      mockedFetch
+    );
+
+    expect(token).toBeNull();
+  });
+});
+
+describe("boundFetch", () => {
+  it("returns a fetch function which adds the access token to requests", async () => {
+    const accessToken = "access-token";
+    const fetchFn = boundFetch(accessToken);
+
+    const mockedFetch = mockFetch(new Response(JSON.stringify({})));
+
+    await fetchFn("https://fake.url");
+
+    expect(mockedFetch).toHaveBeenCalledWith("https://fake.url", {
+      headers: {
+        authorization: `Bearer ${accessToken}`,
+      },
+    });
+  });
+
+  it("merges initalization options in the fetch fn", async () => {
+    const accessToken = "access-token";
+    const fetchFn = boundFetch(accessToken);
+
+    const mockedFetch = mockFetch(new Response(JSON.stringify({})));
+
+    await fetchFn("https://fake.url", {
+      method: "POST",
+      headers: {
+        "x-some-header": "some-value",
+      },
+    });
+
+    expect(mockedFetch).toHaveBeenCalledWith("https://fake.url", {
+      method: "POST",
+      headers: {
+        authorization: `Bearer ${accessToken}`,
+        "x-some-header": "some-value",
+      },
+    });
+  });
+});
+
+describe("fetchWithVc", () => {
+  // These tests may be fairly brittle due to the nature of the mocked calls.
+  it("throws an error if no www-authentication header is found", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+
+    mockFetch(new Response());
+
+    await expect(fetchWithVc(resourceIri, MOCK_VC)).rejects.toThrow(
+      "No www-authentication header found"
+    );
+  });
+
+  it("throws an error if www-authentication header does not contain a ticket", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+
+    mockFetch(
+      new Response(undefined, {
+        headers: {
+          "WWW-Authenticate": 'UMA realm="test" as_uri="https://fake.url"',
+        },
+      })
+    );
+
+    await expect(fetchWithVc(resourceIri, MOCK_VC)).rejects.toThrow(
+      'did not include "ticket"'
+    );
+  });
+
+  it("throws an error if www-authentication header does not contain an uma endpoint iri", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+
+    mockFetch(
+      new Response(undefined, {
+        headers: {
+          "WWW-Authenticate": 'UMA realm="test" ticket="some_value"',
+        },
+      })
+    );
+
+    await expect(fetchWithVc(resourceIri, MOCK_VC)).rejects.toThrow(
+      'did not include "as_uri"'
+    );
+  });
+
+  it("gets uma configuration from the iri in the auth header", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+    const mockHeader =
+      'UMA realm="test" as_uri="https://fake.url" ticket="some_value"';
+
+    (crossFetch as jest.MockedFunction<typeof crossFetch>)
+      // first request is for headers
+      .mockResolvedValueOnce(
+        new Response(undefined, {
+          headers: {
+            "WWW-Authenticate": mockHeader,
+          },
+        })
+      )
+      // second request is for config
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            token_endpoint: "https://fake.url/token-endpoint",
+          })
+        )
+      )
+      // third request is for the token
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            access_token: "access-token",
+          })
+        )
+      );
+
+    await fetchWithVc(resourceIri, MOCK_VC);
+
+    expect(crossFetch).toHaveBeenCalledWith(
+      "https://fake.url/.well-known/uma2-configuration"
+    );
+  });
+
+  it("throws an error if an access token isn't returned", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+    const mockHeader =
+      'UMA realm="test" as_uri="https://fake.url" ticket="some_value"';
+
+    (crossFetch as jest.MockedFunction<typeof crossFetch>)
+      // first request is for headers
+      .mockResolvedValueOnce(
+        new Response(undefined, {
+          headers: {
+            "WWW-Authenticate": mockHeader,
+          },
+        })
+      )
+      // second request is for config
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            token_endpoint: "https://fake.url/token-endpoint",
+          })
+        )
+      )
+      // third request is for the token. Note that the token is missing here.
+      .mockResolvedValueOnce(new Response(JSON.stringify({})));
+
+    await expect(fetchWithVc(resourceIri, MOCK_VC)).rejects.toThrow(
+      "No access token was returned"
+    );
+  });
+
+  it("uses the provided fetch (if any) for the token request", async () => {
+    const resourceIri = "https://fake.url/some-resource";
+    const mockHeader =
+      'UMA realm="test" as_uri="https://fake.url" ticket="some_value"';
+    // Cross-fetch is necessarily called for unauthenticated calls.
+    (crossFetch as jest.MockedFunction<typeof crossFetch>)
+      // first request is for headers
+      .mockResolvedValueOnce(
+        new Response(undefined, {
+          headers: {
+            "WWW-Authenticate": mockHeader,
+          },
+        })
+      )
+      // second request is for config
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            token_endpoint: "https://fake.url/token-endpoint",
+          })
+        )
+      );
+    const mockedFetch = jest
+      .fn(global.fetch)
+      // third request is for the token
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            access_token: "access-token",
+          })
+        )
+      );
+
+    await fetchWithVc(resourceIri, MOCK_VC, { fetch: mockedFetch });
+
+    expect(mockedFetch).toHaveBeenCalled();
+  });
+});