+++ src/resource/wrapSolidClientFunction.ts
@@ -1,0 +1,33 @@
+// Copyright 2021 Inrupt Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+// Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+// PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+// SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// We just want the generic type to be callable, the types of its
+// arguments and return value are inferred using TS utility types,
+// so it's fine to have any in the generic.
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+
+/* eslint @typescript-eslint/no-explicit-any: 0 */
+export async function wrapSolidClientFunction<T extends (...args: any) => any>(
+  solidClientFunction: T,
+  wrappedParameters: Parameters<T>,
+  options?: { fetch: typeof global.fetch }
+): Promise<ReturnType<T>> {
+  // We know that all the functions we want to wrap accept an options object
+  // with a fetch field as their last argument.
+  return solidClientFunction(...wrappedParameters.values(), options);
+}